Code cleanup
General code cleanup to fix compiler warnings.
disabled midi input due to initialization crashing in one of the test

Fixes https://github.com/4rknova/xtracer/issues/54

+++ src/frontend/gui/graphics.cc
@@ -37,7 +37,7 @@ void draw(NSVGimage *image, float position_x, float position_y)
 
     for (NSVGshape *shape = image->shapes; shape != NULL; shape = shape->next) {
 	    for (NSVGpath *path = shape->paths; path != NULL; path = path->next) {
-    		for (size_t i = 0; i < path->npts-1; i += 3) {
+    		for (int i = 0; i < path->npts-1; i += 3) {
 	    		float* p = &path->pts[i*2];
 
                 ImColor col = ImColor(shape->stroke.color);
